+++ packages/plugin-equation/src/ElementEquationInline/ElementEquationInline.tsx
@@ -1,13 +1,7 @@
 import React, { Fragment, useState, useEffect, FormEvent, useRef } from 'react';
-import {
-  RenderElementProps,
-  useEditor,
-  useFocused,
-  useSelected,
-  ReactEditor,
-} from 'slate-react';
+import { useEditor, useFocused, useSelected, ReactEditor } from 'slate-react';
 import { Transforms, Element } from 'slate';
-import { useUI, useHardReturn } from '@slash/editor';
+import { useUI, useHardReturn, RenderElementProps } from '@slash/editor';
 import EquationTextarea from '../EquationTextarea';
 import useTex from '../utils/useTex';
 import { EquationIcon, EnterIcon } from '../icons';