+++ packages/plugin-heading/src/ElementHeadingThree/ElementHeadingThree.test.tsx
@@ -8,8 +8,12 @@ const TYPE = 'foo';
 const TEXT = 'Some text';
 
 const defaultProps: ElementHeadingThreeProps = {
-  attributes: { 'data-slate-node': 'element', ref: React.createRef() },
-  element: { type: TYPE, children: [{ text: TEXT }] },
+  attributes: {
+    'data-slate-node': 'element',
+    'data-block-id': 'block-1',
+    ref: React.createRef(),
+  },
+  element: { type: TYPE, children: [{ text: TEXT }], id: 'block-id' },
   children: <div>{TEXT}</div>,
 };