+++ packages/plugin-heading/CHANGELOG.md
@@ -1,7 +1,19 @@
 # Change Log
 
-All notable changes to this package will be documented in this file.
-See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+This is a list of changes to plugin-heading with each new release. Until 1.0.0 is released, breaking changes will be added as minor or patch version bumps.
+
+## 0.1.0
+
+### `0.0.1` — August 3, 2020
+
+###### BREAKING CHANGE
+
+- Update to new RenderElementProps API
 
 ## 0.0.1
-###### *July 30, 2020*
+
+### `0.0.1` — July 30, 2020
+
+###### NEW
+
+- Created plugin